requirements.txt
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==42.0.8
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.4
django-appconf==1.0.6
django-browser-reload==1.13.0
django-cleanup==8.1.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-htmx==1.18.0
django-tailwind==3.6.0
django_compressor==4.5
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
pillow==10.3.0
psycopg2==2.9.7
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.2
redis==5.0.8
requests==2.32.3
rich==13.7.1
rjsmin==1.2.2
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
text-unidecode==1.3
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
zope.interface==6.4.post2
